Added a minor clarification to the wse SPEC.md. Completed the tests for escaped text encoding. Moved a couple of test classes to make the package structure more logical.
