+++ specification/wse/src/main/scripts/org/kaazing/specification/wse/text.escaped.encoding/echo.binary.with.fragmented.encoded.byte/request.rpt
@@ -15,10 +15,8 @@
 #
 
 property sequence ${wse:randomInt(100)}
-property allPossibleBytes ${wse:allBytes()}
-property utf8AllBytes ${wse:encodeBytesAsUtf8(allPossibleBytes)}
 
-connect http://localhost:8080/path/;e/ctm?query
+connect http://localhost:8080/path/;e/ctem?query
 connected
 
 write method "POST"
@@ -54,8 +52,8 @@ read status "200" /.+/
 read header "Content-Type" "text/plain;charset=windows-1252"
 read header "Connection" "close"
 
-read [0x80 0x82 0x00]
-read ${allPossibleBytes}
+read [0x80 0x01 0x84]
+
 
 # Upstream
 connect await CREATED
@@ -67,10 +65,9 @@ write version "HTTP/1.1"
 write header host
 write header "Content-Type" "text/plain;charset=utf-8"
 write header "X-Sequence-No" ${wse:asString(sequence + 1)}
-write header content-length
+write header "content-length" "5"
 
-# Unencoded form is 0x80 0x82 0x00
-# Frame length is 256 (logical number of bytes)
-write [0xC2 0x80 0xC2 0x82 0x00] ${utf8AllBytes}
-write flush
+write [0xC2 0x80 0x01 0xC2]
+write await WRITE_SECOND_FRAGMENT
+write [0x84]
