+++ specification/wse/src/main/scripts/org/kaazing/specification/wse/text.escaped.encoding/echo.binary.with.fragmented.escaped.byte/request.rpt
@@ -52,7 +52,7 @@ read status "200" /.+/
 read header "Content-Type" "text/plain;charset=windows-1252"
 read header "Connection" "close"
 
-read [0x80 0x02 0x10 0x11]
+read [0x80 0x01 0x7f 0x6e]
 
 
 # Upstream
@@ -65,8 +65,9 @@ write version "HTTP/1.1"
 write header host
 write header "Content-Type" "text/plain;charset=utf-8"
 write header "X-Sequence-No" ${wse:asString(sequence + 1)}
-write header content-length
+write header "content-length" "5"
 
-write [0xC2 0x80 0x02 0x10 0x11]
-write flush
+write [0xC2 0x80 0x01 0x7f]
+write await WRITE_SECOND_FRAGMENT
+write [0x6e]
