+++ specification/wse/src/main/scripts/org/kaazing/specification/wse/text.escaped.encoding/echo.binary.with.fragmented.escaped.byte/response.rpt
@@ -19,10 +19,7 @@ property downstreamId ${wse:uniqueId()}
 property upstream ${wse:append("http://localhost:8080/path/", upstreamId)}
 property downstream ${wse:append("http://localhost:8080/path/", downstreamId)}
 
-property allPossibleBytes ${wse:allBytes()}
-property utf8AllBytes ${wse:encodeBytesAsUtf8(allPossibleBytes)}
-
-accept http://localhost:8080/path/;e/ctm?query
+accept http://localhost:8080/path/;e/ctem?query
 accepted
 connected
 
@@ -60,9 +57,8 @@ write header "Connection" "close"
 write flush
 
 write await DATA_RECEIVED
-write [0x80 0x82 0x00]
-write ${allPossibleBytes}
-write flush
+write [0x80 0x01 0x7f 0x6e]
+
 
 # Upstream
 accept ${upstream}
@@ -75,8 +71,5 @@ read version "HTTP/1.1"
 read header "Content-Type" /text\/plain;charset=(?i)utf-8/
 read header "X-Sequence-No" ${wse:asString(sequence + 1)}
 
-read [0xC2 0x80 0xC2 0x82 0x00]
-# Encoded data length is 3 * 128 because half the bytes have leading bit set so need 2 bytes in UTF-8
-read ${utf8AllBytes}
+read [0xC2 0x80 0x01 0x7f 0x6e]
 read notify DATA_RECEIVED
-
