+++ specification/wse/src/test/java/org/kaazing/specification/wse/internal/EncodingTest.java
@@ -13,11 +13,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.kaazing.specification.wse;
+package org.kaazing.specification.wse.internal;
 
 import org.junit.Assert;
 import org.junit.Test;
-import org.kaazing.specification.wse.internal.Encoding;
 
 public class EncodingTest {
